+++ src/WebJobs.Extensions.DurableTask/ContextImplementations/DurableActivityContext.cs
@@ -20,6 +20,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
         private readonly string instanceId;
         private readonly MessagePayloadDataConverter messageDataConverter;
         private readonly bool inputsAreArrays;
+        private readonly bool rawInput;
 
         private JToken parsedJsonInput;
         private string serializedOutput;
@@ -35,6 +36,9 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
             this.inputsAreArrays =
                 config.OutOfProcProtocol == OutOfProcOrchestrationProtocol.OrchestratorShim ||
                 config.PlatformInformationService.GetWorkerRuntimeType() == WorkerRuntimeType.DotNetIsolated;
+
+            // Do not manipulate JSON input when using middleware passthrough.
+            this.rawInput = config.OutOfProcProtocol == OutOfProcOrchestrationProtocol.MiddlewarePassthrough;
         }
 
         /// <inheritdoc />
@@ -116,21 +120,23 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
                 return null;
             }
 
-            var value = jToken as JValue;
-            if (value != null)
+            if (!this.rawInput && jToken is JValue value)
             {
                 return value.ToObject(destinationType);
             }
 
             string serializedValue = jToken.ToString(Formatting.None);
 
-            // Object inputs for out-of-proc activities are passed in their JSON-stringified form with a destination
-            // type of System.String. Unfortunately, deserializing a JSON string to a string causes
-            // MessagePayloadDataConverter to throw an exception. This is a workaround for that case. All other
-            // inputs with destination System.String (in-proc: JSON and not JSON; out-of-proc: not-JSON) inputs with
-            // destination System.String should cast to JValues and be handled above.)
-            if (destinationType.Equals(typeof(string)))
+            if (this.rawInput) // the "modern" OOProc protocol case
             {
+                return serializedValue;
+            }
+            else if (destinationType.Equals(typeof(string))) // legacy OOProc protocol case (JS, Python, PowerShell)
+            {
+                // Object/complex inputs in "legacy" out-of-proc activities are passed with a destination
+                // type of System.String (to be precise, if inspected with a debugger, you'll see a stringified JSON).
+                // Unfortunately, deserializing a JSON string to a string causes  MessagePayloadDataConverter to throw an exception, so the
+                // return statement prevents that.
                 return serializedValue;
             }