+++ src/WebJobs.Extensions.DurableTask/ContextImplementations/RemoteOrchestratorContext.cs
@@ -5,6 +5,8 @@ using System;
 using System.Collections.Generic;
 using DurableTask.Core;
 using DurableTask.Core.Command;
+using DurableTask.Core.Entities;
+using DurableTask.Core.Exceptions;
 using DurableTask.Core.History;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
@@ -17,9 +19,12 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
 
         private OrchestratorExecutionResult? executionResult;
 
-        public RemoteOrchestratorContext(OrchestrationRuntimeState runtimeState)
+        private Exception? failure;
+
+        public RemoteOrchestratorContext(OrchestrationRuntimeState runtimeState, TaskOrchestrationEntityParameters? entityParameters)
         {
             this.runtimeState = runtimeState ?? throw new ArgumentNullException(nameof(runtimeState));
+            this.EntityParameters = entityParameters;
         }
 
         [JsonProperty("instanceId")]
@@ -42,6 +47,36 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
 
         [JsonIgnore]
         internal string? SerializedOutput { get; private set; }
+
+        [JsonIgnore]
+        internal TaskOrchestrationEntityParameters? EntityParameters { get; private set; }
+
+        internal void ThrowIfFailed()
+        {
+            if (this.failure != null)
+            {
+                throw this.failure;
+            }
+        }
+
+        internal OrchestratorExecutionResult GetResult()
+        {
+            return this.executionResult ?? throw new InvalidOperationException($"The execution result has not yet been set using {nameof(this.SetResult)}.");
+        }
+
+        internal bool TryGetOrchestrationErrorDetails(out string details)
+        {
+            if (this.failure != null)
+            {
+                details = this.failure.Message;
+                return true;
+            }
+            else
+            {
+                details = string.Empty;
+                return false;
+            }
+        }
 
         internal void SetResult(IEnumerable<OrchestratorAction> actions, string customStatus)
         {
@@ -102,16 +137,24 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask
                     this.OrchestratorCompleted = true;
                     this.SerializedOutput = completeAction.Result;
                     this.ContinuedAsNew = completeAction.OrchestrationStatus == OrchestrationStatus.ContinuedAsNew;
+
+                    if (completeAction.OrchestrationStatus == OrchestrationStatus.Failed)
+                    {
+                        string message = completeAction switch
+                        {
+                            { FailureDetails: { } f } => f.ErrorMessage,
+                            { Result: { } r } => r,
+                            _ => "Exception occurred during orchestration execution.",
+                        };
+
+                        this.failure = new OrchestrationFailureException(message);
+                    }
+
                     break;
                 }
             }
 
             this.executionResult = result;
-        }
-
-        internal OrchestratorExecutionResult GetResult()
-        {
-            return this.executionResult ?? throw new InvalidOperationException($"The execution result has not yet been set using {nameof(this.SetResult)}.");
         }
     }
 }