+++ release_notes.md
@@ -1,5 +1,20 @@
 # Release Notes
 
+## Microsoft.Azure.Functions.Worker.Extensions.DurableTask 1.2.0
+
+### New Features
+
+- Add `suspendPostUri` and `resumePostUri` to the list of returned URIs in `CreateCheckStatusResponseAsync`. (https://github.com/Azure/azure-functions-durable-extension/pull/2785)
+- Fix `NotSupportedException` when calling `PurgeAllInstancesAsync` and `PurgeInstanceAsync`
+
+### Bug Fixes
+
+### Breaking Changes
+
+### Dependency Updates
+
+## Microsoft.Azure.WebJobs.Extensions.DurableTask <version>
+
 ### New Features
 
 ### Bug Fixes